requirements.txt
flask
flask-cors
tensorflow
numpy
Pillow
google-generativeai



